main utils file - move form related helpers to a separate file
